Use file to load the test
